src/main/webapp/app/entities/job-sub-cat/delete/job-sub-cat-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { JobSubCatService } from '../service/job-sub-cat.service';

import { JobSubCatDeleteDialogComponent } from './job-sub-cat-delete-dialog.component';

describe('JobSubCat Management Delete Component', () => {
  let comp: JobSubCatDeleteDialogComponent;
  let fixture: ComponentFixture<JobSubCatDeleteDialogComponent>;
  let service: JobSubCatService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      declarations: [JobSubCatDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(JobSubCatDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(JobSubCatDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(JobSubCatService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      })
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
